Removed condense(), added aggregateIntransitive()
I added aggregateIntransitive(), which does the same thing as
aggregateTransitive but doesn't assume that the criterion is transitive.

I unified the parameters for both aggregate functions into an
AggregationConfig interface.

I removed the condense() function because aggregateIntransitive will do
exactly what it does. I'll probably come up with some default criteria
callbacks that people may want to use often.
