+++ src/utilities.ts
@@ -4,6 +4,9 @@ export {
   generateSequenceAnnotations,
   GenerationPattern,
 } from "./utilities/annotation-generation";
-export { aggregateTransitive, condenseAnnotations } from "./utilities/condense";
+export {
+  aggregateTransitive,
+  aggregateIntransitive,
+} from "./utilities/aggregate";
 export { exportPng } from "./utilities/export-png";
 export { generateId } from "./utilities/id-generation";
